COMP: Mac 10.15 XCode 11.4.1 build of openssl
CMake computes the value of the -isysroot flag for cmake builds.
OpenSSL does not have a mechanism for setting this directly, so
using an environmental variable allows the compiler to pick
up the correct SDKROOT path.

Allow the SDKROOT to be found on mac when building OpenSSL.
This is needed to find standard header files and avoid errors like

```
 ./cryptlib.h:62:11: fatal error: 'stdlib.h' file not found
 # include <stdlib.h>
```