COMP: Add QT version check for tabWidth usage in qMRMLNodeComboBoxMenuDelegate
In Qt 5, QStyleOptionMenuItem had a tabWidth member that controlled the space reserved for a tab character (e.g., for aligning shortcut text). It was often set to 0 to disable tab-based spacing.

In Qt 6, tabWidth was removed as part of a general cleanup of the style options API, since it’s no longer used by the style engine.